+++ CHANGELOG.md
@@ -4,11 +4,11 @@
 
 -   Add untyped constructor for `PolarizedMap` ([PR#49](https://github.com/ziotom78/Healpix.jl/pull/49))
 
--   Implement `udgrade` ([PR#46](https://github.com/ziotom78/Healpix.jl/pull/46))
+-   Implement `udgrade` ([PR#46](https://github.com/ziotom78/Healpix.jl/pull/46), [#51](https://github.com/ziotom78/Healpix.jl/pull/51))
 
 -   Implement `pixwin` ([PR#45](https://github.com/ziotom78/Healpix.jl/pull/45))
 
--   Implement `applyweights!` ([PR#41](https://github.com/ziotom78/Healpix.jl/pull/41) and [#44](https://github.com/ziotom78/Healpix.jl/pull/44))
+-   Implement `applyfullweights!` ([PR#41](https://github.com/ziotom78/Healpix.jl/pull/41), [#44](https://github.com/ziotom78/Healpix.jl/pull/44)), and [#52](https://github.com/ziotom78/Healpix.jl/pull/52)
 
 -   Implement `ring2nest!` and `nest2ring!` ([PR#40](https://github.com/ziotom78/Healpix.jl/pull/40))